requirements.txt
bleach==1.5.0
certifi==2017.7.27.1
chardet==3.0.4
click==6.7
cycler==0.10.0
Flask==0.12.2
html5lib==0.9999999
idna==2.6
itsdangerous==0.24
Jinja2==2.9.6
Markdown==2.6.9
MarkupSafe==1.0
matplotlib==2.0.2
numpy==1.13.1
olefile==0.44
Pillow==4.2.1
protobuf==3.4.0
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2017.2
requests==2.18.4
six==1.10.0
tensorflow==1.3.0
tensorflow-tensorboard==0.1.4
urllib3==1.22
Werkzeug==0.12.2
